post_service/requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pytest==7.4.4
python-dateutil==2.8.2